BACKEND/requirements.txt
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
DateTime==4.3
ecdsa==0.14.1
fastapi==0.61.1
fastapi-utils==0.2.1
flake8==3.8.4
h11==0.10.0
idna==2.10
mccabe==0.6.1
mysql-connector==2.2.9
passlib==1.7.2
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.6.1
pyflakes==2.2.0
PyMySQL==0.10.1
pyserial==3.4
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.6
typing==3.7.4.3
urllib3==1.25.10
uvicorn==0.12.1
zope.interface==5.1.1
